renaming errors fixed
