+++ front-end/src/App.tsx
@@ -1,8 +1,8 @@
 import React, { Suspense } from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import Home from './pages/home/Home'
-import SignIn from './pages/signIn/SignIn'
-import SignUp from './pages/signUp/SignUp'
+import Home from './pages/Home/Home'
+import SignIn from './pages/SignIn/SignIn'
+import SignUp from './pages/SignUp/SignUp'
 import GuardRoute from './services/GuardRoute'
 
 const App = () => {
