+++ front-end/src/pages/SignUp/SignUp.tsx
@@ -16,7 +16,7 @@ import Visibility from '@mui/icons-material/Visibility'
 import VisibilityOff from '@mui/icons-material/VisibilityOff'
 import { useNavigate } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux/es/exports'
-import { addUser, setAddStatus } from '../../slices/userSlice'
+import { addUser, setAddStatus } from '../../slices/UserSlice'
 import { User } from '../../utils/interface'
 import LoadingSpinner from '../../components/loadingSpinner/LoadingSpinner'
